Add support for generating bitsets from proc macro
These are a bit funny, and our DSL syntax is getting a bit weird:
basically this switches behaviour based on the presence of a #[flags(type)]
annotation at the top-level.

We use the bitflags crate to generate the flags, and we also generate an
impl of the Scalar trait, so that these types can actually be used
directly in another type definition, as in you can have a
`BigEnding<Flags>` field. Will this end up being useful? We shall see...