[codegen] Compute NeededWhen union
This fixes a bug where, if a field was needed by multiple other fields
in different contexts, the last reference would overwrite any previous
references.

With this patch, we check if there is a previous reference, and compute
the 'union of need' as necessary.

This will come up in the CPAL table.