[ci] Ensure codegen is up-to-date
This adds a github actions task that runs the codegen tool and ensures
it does not produce any changes; this ensures that generated files which
have been commited match the commited inputs.

This might cause issues when the files have been generated in windows?
If this is the case we should figure out why, and get windows producing
the same outputs as the unixes.