[codegen] Manually reformat generated bitflags! decls
Because the contents of this macro is not value Rust code, rustfmt
doesn't know what to do with it, which means we currently end up with
monolithic single line blobs in our generated files for any declared
bitflags.

This patch adds an additional post-processing step that finds any such
declarations and manually reparses and rewrites them.

Maybe this is not a good idea? This was definitely me scratching an itch
at the end of the day.