Add font-test-data crate
This adds a new crate, which hosts the raw test data that is shared
between the other crates.

This lets us remove the 'test_data' feature from read-fonts, which was
always a bit of a hack.

This also adds an Into<FontData<'a>> for &'a [u8], which makes it easier
to instantiate an object from raw bytes, which is useful because the new
crate cannot expose FontData directly.