[read-fonts] CFF charstring evaluation
Adds support for charstring evaluation. Results are passed to an impl of the `CommandSink` trait that provides hooks for processing both path commands and hinting data.

This doesn't do scaling or hinting, but both can be built on top of the sink trait.

Removes the existing "low level" parsing code for now because that did not properly handle the hint/counter masks args in the data stream that depend on tracking the stem hint count (and thus, also recursing into subroutines which can apply stem hints). I'll think about how to properly address this.

Testing is very light. We need a bit more infra for building a full PostScript "font" object from the CFF/CFF2 tables to support more sophisticated evaluation. Opened #478 to track testing.