[font-types] Remove panicking paths from Tag
This was ill-considered, and is not a good match for the ways this type
is used in practice.

More broadly, this change represents a loosening of constraints on the
contents of a Tag, to better match what we are actually doing anyway.
Concretely, this means that we are taking the spec as a *suggestion*,
but fully support invalid tags.

This also adds a `validate` method to Tag, which can be used to check
the validity of a tag that was created through one of the unchecked
paths.