[iup] Fix mismatch in behaviour with fonttools
This patch fixes an issue with iup not resolving, although I do not
fully understand the problem.

Basically: in comparing our implementation with the one in fonttools,
there is clear divergence in some of the numerical code, where we are
using Option<usize> in places where python is just using an integer.
More importantly, we are using None where Python is using -1, but our
comparison code is never checking if a value == None, whereas python
*does* check if a value == -1, which means that python is setting the
solution in cases where we are not.

This can be fixed by using `checked_sub` instead of `saturating_sub`, at
least for this input.

It feels like this is probably fine, since the structure of this loop is
such that the only negative value that is possible is -1 (at the start)
but it feels a little loose.