[write-fonts] Add RemapVariationIndices trait
After compiling an ItemVariationStore, we need to remap any existing
indices into that structure, in any place that they exist.

This trait encapsulates that behaviour.

It is currently only implemented by GPOS, but only because I'm not very
familiar with its use elsewhere, and I want to make sure this works.