[read-fonts] Only inspect reachable lookups for closure
Previously we just looked at every lookup in the lookup list when
computing the closure. This sounds okay at first glance, but it
overlooks the fact that the lookup list can include lookups that are
only reachable via other contextual lookups.

This introduces a much more rigorous approach, where on each run we
recalculate the set of reachable lookups, by checking to see if the
current set of glyphs matches any (chain) contextual rules, and then
adding the appropriate lookups from those rules.