[skrifa] Do not issue COLRv1 draw command for empty sweep (#1073)
If even after normalization the angles match and the sector is empty,
do not issue a draw command. The spec has an extra sentence describing
that situation. Fixes a difference between FreeType and Fontations
COLRv1 backends in Skia as well, see GM_colrv1_sweep_coincident gm
test. Covered by that test and the sweep-coincident glyphs in the COLRv1
test font.

Fixes #1017.