[skrifa] new hinting options type (#1087)
Replaces the current HintingMode with a new HintingOptions type to support autohinting. Specifically, it allows for selecting the hinting engine (interpreter, auto or auto fallback). Also reintroduces a "light" mode to match FT_RENDER_MODE_LIGHT since this actually represents a different mode in the autohinter (but didn't for TrueType or CFF).

Adds the symmetric_rendering flag as well (closes #1080).