Only write explicit tuple min/maxes if there is an intermediate tent (#1345)
* Only write explicit tuple min/maxes if there is an intermediate tent

Moved from original implementation in googlefonts/fontc#1225, see
discussion.

This optimisation is also implemented in fontTools, and reduced `gvar`
table size by ~50% when testing locally on a large variable font.

* Write intermediates if _any_ are required, not if _all_ are required

Unfortunately we lose some of the simplicity of the prior approach, as
there is no Option `collect()` for `any()`, and we need to make the
implied intermediates available outside of the coordinates struct.

* Fix spelling after refactor

* Test logic for only serialising intermediates that cannot be implied

* Improve readability & brevity

* Document new public types and functions

* Correct typo

* [gvar] Sketch out simplified GlyphDeltas API

Fewer types and fewer code paths.

---------

Co-authored-by: Ricky Atkins <ricky.atkins@daltonmaag.com>
Co-authored-by: Colin Rofls <colin@cmyr.net>