Fix CFF scaling with a nested font matrix (#1644)
We recently added support for font matrices in a top dict and this patch extends that to support font matrices in a font dict. This goes a bit further in that it now supports font matrices in both dicts, concatenating and scaling as required. An additional test is included to ensure we match FreeType for this behavior.

ref #1638